Factor Integration

+++ src/sleepAppDatabase/Database.java
@@ -274,7 +274,7 @@ public class Database {
         }
     }
 
-    public static boolean addScreenTimeEntry(double screentime, int day, int month, int year){
+    public static boolean addScreenTimeEntry(int screentime, int day, int month, int year){
         Date addDate = Date.valueOf(year + "-" + month + "-" + day);
         try {
             Connection conn = DriverManager.getConnection(databaseURL);
